feature/tags modify components files of tag

+++ src/modules/tags/application/use-cases/delete-tag.use-case.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { LoggingService } from 'src/modules/shared/logging/domain/services/logging.service';
-
+import { TagRepository } from '../../domain/repository/tag.repository';
+import { UserEntity } from 'src/modules/users/domain/entities/user.entity';
+import { UserCannotDeleteTagException } from '../../domain/exceptions/user-cannot-delete-tag.exception';
 @Injectable()
 export class DeleteTagUseCase {
   constructor(
@@ -8,7 +10,11 @@ export class DeleteTagUseCase {
     private readonly loggingService: LoggingService,
   ) {}
 
-  public async execute(id: string): Promise<void> {
+  public async execute(id: string, user: UserEntity): Promise<void> {
+    if (!user.permissions.tags.canDelete()) {
+      throw new UserCannotDeleteTagException();
+    }
+
     this.loggingService.log('DeleteTagUseCase.execute');
     await this.tagRepository.deleteTag(id);
   }
